EES-1319 Refactor out separate ChartDataSetsAddForm from ChartDataSetsConfiguration
