+++ src/explore-education-statistics-admin/src/pages/release/datablocks/components/ChartDataSetsConfiguration.module.scss
@@ -1,0 +1,20 @@
+@import '~govuk-frontend/govuk/base';
+
+.dataSets {
+  list-style: none;
+  padding-left: 0;
+
+  li {
+    border-bottom: 1px solid govuk-colour('mid-grey');
+    margin-bottom: govuk-spacing(4);
+  }
+
+  &:last-of-type {
+    border-bottom: 0;
+  }
+}
+
+.dataSetRow {
+  display: flex;
+  justify-content: space-between;
+}
